+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Search: disable domain button for autosuggest results
 - Timeline: fix inconsistent padding
 - UI: fixed legacy domain button rendered too big
+- Profile: fixed 'Follows you' badge rendered as 'INVALID!'
 
 ## 2023.10.4
 - UX: fetch account when mention component is rendering